--- azext_edge/edge/providers/rpsaas/adr/common.py
@@ -1,17 +1,0 @@
-# coding=utf-8
-# ----------------------------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License file in the project root for license information.
-# ----------------------------------------------------------------------------------------------
-
-from enum import Enum
-
-
-class IdentityType(Enum):
-    """
-    Types of managed identity for ARM resources.
-    """
-    system_assigned = "SystemAssigned"
-    user_assigned = "UserAssigned"
-    system_assigned_user_assigned = "SystemAssigned, UserAssigned"
-    none = "None"
